Refactor menus
I didn't like the fact that menus were globally registered long lived
objects whose evaluated items had to be re(set) on every rendering. Not
sure if that was likely to cause any issue (concurrency, maybe) but I
still like it better to separate long lived globally registered objects
(`MenuRegistry` now) from each instance of an actual menu dependent on
the view context (`Menu` now).
