+++ decidim-core/app/helpers/decidim/menu_helper.rb
@@ -5,7 +5,7 @@ module Decidim
   module MenuHelper
     # Public: Returns the main menu presenter object
     def main_menu
-      @main_menu ||= MenuPresenter.new(Menu.find(:menu), self)
+      @main_menu ||= MenuPresenter.new(:menu, self)
     end
   end
 end
